backend/pyproject.toml
[project]
name = "vece-backend"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosmtplib>=4.0.0",
    "aiosqlite>=0.21.0",
    "asyncpg>=0.30.0",
    "python-dotenv>=1.0.0",
    "boto3>=1.37.1",
    "faker>=37.1.0",
    "google-generativeai>=0.8.0",
    "fastapi[standard]>=0.115.8",
    "mypy>=1.15.0",
    "passlib>=1.7.4",
    "pillow>=11.1.0",
    "pillow-heif>=0.20.0",
    "pip>=25.0.1",
    "pre-commit>=4.1.0",
    "psycopg>=3.2.6",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "ruff>=0.9.7",
    "scikit-learn>=1.6.1",
    "sqlalchemy>=2.0.38",
    "types-passlib>=1.7.7.20250401",
]

[tool.mypy]

[[tool.mypy.overrides]]
module = "sklearn.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
markers = [
    "asyncio: mark test as asyncio"
]
